+++ index.ts
@@ -6,8 +6,8 @@
  * credential classes for programmatic consumers.
  */
 export { ServicelyApi } from './credentials/ServicelyApi.credentials';
-export { ServicelyAIToolAuthApi } from './credentials/ServicelyAIToolAuthApi.credentials';
+export { ServicelySoFiAIWebhookAuthApi } from './credentials/ServicelySoFiAIWebhookAuthApi.credentials';
 export { Servicely } from './nodes/Servicely/Servicely.node';
 export { ServicelyTrigger } from './nodes/Servicely/ServicelyTrigger.node';
-export { ServicelyAIToolTrigger } from './nodes/ServicelyAITool/ServicelyAIToolTrigger.node';
+export { ServicelySoFiAIWebhookTrigger } from './nodes/ServicelySoFiAIWebhook/ServicelySoFiAIWebhookTrigger.node';
 export * from './nodes/Servicely/types';